requirements.txt
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
firebase-admin==2.17.0
Flask==1.0.3
google-api-core==1.11.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.2
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
twilio==6.27.1
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
